requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
diffusers==0.24.0
accelerate==0.25.0
datasets==2.15.0
nltk==3.8.1
opencv-python==4.8.1.78
matplotlib==3.8.2
seaborn==0.13.0
wandb==0.16.1
tensorboard==2.15.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1